requirements.txt
streamlit==1.32.0
pandas==2.0.3
numpy==1.24.3
mlxtend==0.22.0
plotly==5.18.0
networkx==3.1
scikit-learn==1.0.2
scipy==1.7.3
statsmodels==0.13.2
reportlab==4.1.0
python-docx==1.1.0
openpyxl==3.1.2
apscheduler==3.10.4
python-dotenv==1.0.1 